Refactor logbook data fetch logic into reusable class
