+++ src/panels/logbook/ha-panel-logbook.ts
@@ -12,8 +12,6 @@ import {
 } from "date-fns";
 import { css, html, LitElement, PropertyValues } from "lit";
 import { customElement, property, state } from "lit/decorators";
-import { isComponentLoaded } from "../../common/config/is_component_loaded";
-import { computeStateDomain } from "../../common/entity/compute_state_domain";
 import { navigate } from "../../common/navigate";
 import {
   createSearchParam,
@@ -21,19 +19,10 @@ import {
 } from "../../common/url/search-params";
 import { computeRTL } from "../../common/util/compute_rtl";
 import "../../components/entity/ha-entity-picker";
-import "../../components/ha-circular-progress";
 import "../../components/ha-date-range-picker";
 import type { DateRangePickerRanges } from "../../components/ha-date-range-picker";
 import "../../components/ha-icon-button";
 import "../../components/ha-menu-button";
-import {
-  clearLogbookCache,
-  getLogbookData,
-  LogbookEntry,
-} from "../../data/logbook";
-import { loadTraceContexts, TraceContexts } from "../../data/trace";
-import { fetchUsers } from "../../data/user";
-import { showAlertDialog } from "../../dialogs/generic/show-dialog-box";
 import "../../layouts/ha-app-layout";
 import { haStyle } from "../../resources/styles";
 import { HomeAssistant } from "../../types";
@@ -45,36 +34,24 @@ export class HaPanelLogbook extends LitElement {
 
   @property({ reflect: true, type: Boolean }) narrow!: boolean;
 
-  @property() _startDate: Date;
+  @state() _time: { range: [Date, Date] };
 
-  @property() _endDate: Date;
-
-  @property() _entityId = "";
-
-  @property() _isLoading = false;
-
-  @property() _entries: LogbookEntry[] = [];
+  @state() _entityId = "";
 
   @property({ reflect: true, type: Boolean }) rtl = false;
 
   @state() private _ranges?: DateRangePickerRanges;
-
-  private _fetchUserPromise?: Promise<void>;
-
-  @state() private _userIdToName = {};
-
-  @state() private _traceContexts: TraceContexts = {};
 
   public constructor() {
     super();
 
     const start = new Date();
     start.setHours(start.getHours() - 2, 0, 0, 0);
-    this._startDate = start;
 
     const end = new Date();
     end.setHours(end.getHours() + 1, 0, 0, 0);
-    this._endDate = end;
+
+    this._time = { range: [start, end] };
   }
 
   protected render() {
@@ -91,19 +68,15 @@ export class HaPanelLogbook extends LitElement {
               @click=${this._refreshLogbook}
               .path=${mdiRefresh}
               .label=${this.hass!.localize("ui.common.refresh")}
-              .disabled=${this._isLoading}
             ></ha-icon-button>
           </app-toolbar>
         </app-header>
 
-        ${this._isLoading ? html`` : ""}
-
         <div class="filters">
           <ha-date-range-picker
             .hass=${this.hass}
-            ?disabled=${this._isLoading}
-            .startDate=${this._startDate}
-            .endDate=${this._endDate}
+            .startDate=${this._time.range[0]}
+            .endDate=${this._time.range[1]}
             .ranges=${this._ranges}
             @change=${this._dateRangeChanged}
           ></ha-date-range-picker>
@@ -114,29 +87,16 @@ export class HaPanelLogbook extends LitElement {
             .label=${this.hass.localize(
               "ui.components.entity.entity-picker.entity"
             )}
-            .disabled=${this._isLoading}
             @change=${this._entityPicked}
           ></ha-entity-picker>
         </div>
 
-        ${this._isLoading
-          ? html`
-              <div class="progress-wrapper">
-                <ha-circular-progress
-                  active
-                  alt=${this.hass.localize("ui.common.loading")}
-                ></ha-circular-progress>
-              </div>
-            `
-          : html`
-              <ha-logbook
-                .hass=${this.hass}
-                .entries=${this._entries}
-                .userIdToName=${this._userIdToName}
-                .traceContexts=${this._traceContexts}
-                virtualize
-              ></ha-logbook>
-            `}
+        <ha-logbook
+          .hass=${this.hass}
+          .time=${this._time}
+          .entityId=${this._entityId}
+          virtualize
+        ></ha-logbook>
       </ha-app-layout>
     `;
   }
@@ -144,8 +104,6 @@ export class HaPanelLogbook extends LitElement {
   protected firstUpdated(changedProps: PropertyValues) {
     super.firstUpdated(changedProps);
     this.hass.loadBackendTranslation("title");
-
-    this._fetchUserPromise = this._fetchUserNames();
 
     const today = new Date();
     const weekStart = startOfWeek(today);
@@ -167,22 +125,21 @@ export class HaPanelLogbook extends LitElement {
     this._entityId = extractSearchParam("entity_id") ?? "";
 
     const startDate = extractSearchParam("start_date");
-    if (startDate) {
-      this._startDate = new Date(startDate);
-    }
     const endDate = extractSearchParam("end_date");
-    if (endDate) {
-      this._endDate = new Date(endDate);
+
+    if (startDate || endDate) {
+      this._time = {
+        range: [
+          startDate ? new Date(startDate) : this._time.range[0],
+          endDate ? new Date(endDate) : this._time.range[1],
+        ],
+      };
     }
   }
 
   protected updated(changedProps: PropertyValues<this>) {
-    if (
-      changedProps.has("_startDate") ||
-      changedProps.has("_endDate") ||
-      changedProps.has("_entityId")
-    ) {
-      this._getData();
+    if (changedProps.has("_time") || changedProps.has("_entityId")) {
+      this._refreshLogbook();
     }
 
     if (changedProps.has("hass")) {
@@ -193,45 +150,14 @@ export class HaPanelLogbook extends LitElement {
     }
   }
 
-  private async _fetchUserNames() {
-    const userIdToName = {};
-
-    // Start loading users
-    const userProm = this.hass.user?.is_admin && fetchUsers(this.hass);
-
-    // Process persons
-    Object.values(this.hass.states).forEach((entity) => {
-      if (
-        entity.attributes.user_id &&
-        computeStateDomain(entity) === "person"
-      ) {
-        this._userIdToName[entity.attributes.user_id] =
-          entity.attributes.friendly_name;
-      }
-    });
-
-    // Process users
-    if (userProm) {
-      const users = await userProm;
-      for (const user of users) {
-        if (!(user.id in userIdToName)) {
-          userIdToName[user.id] = user.name;
-        }
-      }
-    }
-
-    this._userIdToName = userIdToName;
-  }
-
   private _dateRangeChanged(ev) {
-    this._startDate = ev.detail.startDate;
+    const startDate = ev.detail.startDate;
     const endDate = ev.detail.endDate;
     if (endDate.getHours() === 0 && endDate.getMinutes() === 0) {
       endDate.setDate(endDate.getDate() + 1);
       endDate.setMilliseconds(endDate.getMilliseconds() - 1);
     }
-    this._endDate = endDate;
-
+    this._time = { range: [startDate, endDate] };
     this._updatePath();
   }
 
@@ -242,60 +168,20 @@ export class HaPanelLogbook extends LitElement {
   }
 
   private _updatePath() {
-    const params: Record<string, string> = {};
+    const params: Record<string, string> = {
+      start_date: this._time.range[0].toISOString(),
+      end_date: this._time.range[1].toISOString(),
+    };
 
     if (this._entityId) {
       params.entity_id = this._entityId;
-    }
-
-    if (this._startDate) {
-      params.start_date = this._startDate.toISOString();
-    }
-
-    if (this._endDate) {
-      params.end_date = this._endDate.toISOString();
     }
 
     navigate(`/logbook?${createSearchParam(params)}`, { replace: true });
   }
 
   private _refreshLogbook() {
-    this._entries = [];
-    clearLogbookCache(
-      this._startDate.toISOString(),
-      this._endDate.toISOString()
-    );
-    this._getData();
-  }
-
-  private async _getData() {
-    this._isLoading = true;
-    let entries;
-    let traceContexts;
-
-    try {
-      [entries, traceContexts] = await Promise.all([
-        getLogbookData(
-          this.hass,
-          this._startDate.toISOString(),
-          this._endDate.toISOString(),
-          this._entityId
-        ),
-        isComponentLoaded(this.hass, "trace") && this.hass.user?.is_admin
-          ? loadTraceContexts(this.hass)
-          : {},
-        this._fetchUserPromise,
-      ]);
-    } catch (err: any) {
-      showAlertDialog(this, {
-        title: this.hass.localize("ui.components.logbook.retrieval_error"),
-        text: err.message,
-      });
-    }
-
-    this._entries = entries;
-    this._traceContexts = traceContexts;
-    this._isLoading = false;
+    this.querySelector("ha-logbook")?.refresh();
   }
 
   static get styles() {
@@ -323,13 +209,6 @@ export class HaPanelLogbook extends LitElement {
 
         .progress-wrapper {
           position: relative;
-        }
-
-        ha-circular-progress {
-          position: absolute;
-          left: 50%;
-          top: 50%;
-          transform: translate(-50%, -50%);
         }
 
         .filters {
