+++ src/panels/lovelace/cards/hui-logbook-card.ts
@@ -9,15 +9,11 @@ import {
 import { customElement, property, state } from "lit/decorators";
 import { classMap } from "lit/directives/class-map";
 import { isComponentLoaded } from "../../../common/config/is_component_loaded";
-import { computeStateDomain } from "../../../common/entity/compute_state_domain";
 import { applyThemesOnElement } from "../../../common/dom/apply_themes_on_element";
-import { throttle } from "../../../common/util/throttle";
 import "../../../components/ha-card";
-import "../../../components/ha-circular-progress";
-import { fetchUsers } from "../../../data/user";
-import { getLogbookData, LogbookEntry } from "../../../data/logbook";
 import type { HomeAssistant } from "../../../types";
 import "../../logbook/ha-logbook";
+import type { HaLogbook } from "../../logbook/ha-logbook";
 import { findEntities } from "../common/find-entities";
 import { processConfigEntities } from "../common/process-config-entities";
 import "../components/hui-warning";
@@ -56,21 +52,9 @@ export class HuiLogbookCard extends LitElement implements LovelaceCard {
 
   @state() private _config?: LogbookCardConfig;
 
-  @state() private _logbookEntries?: LogbookEntry[];
+  @state() private _time?: HaLogbook["time"];
 
-  @state() private _configEntities?: EntityConfig[];
-
-  @state() private _userIdToName = {};
-
-  private _lastLogbookDate?: Date;
-
-  private _fetchUserPromise?: Promise<void>;
-
-  private _error?: string;
-
-  private _throttleGetLogbookEntries = throttle(() => {
-    this._getLogBookData();
-  }, 10000);
+  @state() private _entityId?: string[];
 
   public getCardSize(): number {
     return 9 + (this._config?.title ? 1 : 0);
@@ -81,45 +65,16 @@ export class HuiLogbookCard extends LitElement implements LovelaceCard {
       throw new Error("Entities must be specified");
     }
 
-    this._configEntities = processConfigEntities<EntityConfig>(config.entities);
-
     this._config = {
       hours_to_show: 24,
       ...config,
     };
-  }
-
-  protected shouldUpdate(changedProps: PropertyValues): boolean {
-    if (
-      changedProps.has("_config") ||
-      changedProps.has("_persons") ||
-      changedProps.has("_logbookEntries")
-    ) {
-      return true;
-    }
-
-    const oldHass = changedProps.get("hass") as HomeAssistant | undefined;
-
-    if (
-      !this._configEntities ||
-      !oldHass ||
-      oldHass.themes !== this.hass!.themes ||
-      oldHass.locale !== this.hass!.locale
-    ) {
-      return true;
-    }
-
-    for (const entity of this._configEntities) {
-      if (oldHass.states[entity.entity] !== this.hass!.states[entity.entity]) {
-        return true;
-      }
-    }
-
-    return false;
-  }
-
-  protected firstUpdated(): void {
-    this._fetchUserPromise = this._fetchUserNames();
+    this._time = {
+      recent: this._config!.hours_to_show! * 60 * 60 * 1000,
+    };
+    this._entityId = processConfigEntities<EntityConfig>(config.entities).map(
+      (entity) => entity.entity
+    );
   }
 
   protected updated(changedProperties: PropertyValues) {
@@ -138,33 +93,6 @@ export class HuiLogbookCard extends LitElement implements LovelaceCard {
       (configChanged && oldConfig?.theme !== this._config.theme)
     ) {
       applyThemesOnElement(this, this.hass.themes, this._config.theme);
-    }
-
-    if (
-      configChanged &&
-      (oldConfig?.entities !== this._config.entities ||
-        oldConfig?.hours_to_show !== this._config!.hours_to_show)
-    ) {
-      this._logbookEntries = undefined;
-      this._lastLogbookDate = undefined;
-
-      if (!this._configEntities) {
-        return;
-      }
-
-      this._throttleGetLogbookEntries();
-      return;
-    }
-
-    if (
-      oldHass &&
-      this._configEntities!.some(
-        (entity) =>
-          oldHass.states[entity.entity] !== this.hass!.states[entity.entity]
-      )
-    ) {
-      // wait for commit of data (we only account for the default setting of 1 sec)
-      setTimeout(this._throttleGetLogbookEntries, 1000);
     }
   }
 
@@ -189,114 +117,17 @@ export class HuiLogbookCard extends LitElement implements LovelaceCard {
         class=${classMap({ "no-header": !this._config!.title })}
       >
         <div class="content">
-          ${this._error
-            ? html`
-                <div class="no-entries">
-                  ${`${this.hass.localize(
-                    "ui.components.logbook.retrieval_error"
-                  )}: ${this._error}`}
-                </div>
-              `
-            : !this._logbookEntries
-            ? html`
-                <ha-circular-progress
-                  active
-                  alt=${this.hass.localize("ui.common.loading")}
-                ></ha-circular-progress>
-              `
-            : this._logbookEntries.length
-            ? html`
-                <ha-logbook
-                  narrow
-                  relative-time
-                  virtualize
-                  .hass=${this.hass}
-                  .entries=${this._logbookEntries}
-                  .userIdToName=${this._userIdToName}
-                ></ha-logbook>
-              `
-            : html`
-                <div class="no-entries">
-                  ${this.hass.localize(
-                    "ui.components.logbook.entries_not_found"
-                  )}
-                </div>
-              `}
+          <ha-logbook
+            .hass=${this.hass}
+            .time=${this._time}
+            .entityId=${this._entityId}
+            narrow
+            relative-time
+            virtualize
+          ></ha-logbook>
         </div>
       </ha-card>
     `;
-  }
-
-  private async _getLogBookData() {
-    if (
-      !this.hass ||
-      !this._config ||
-      !isComponentLoaded(this.hass, "logbook")
-    ) {
-      return;
-    }
-
-    const hoursToShowDate = new Date(
-      new Date().getTime() - this._config!.hours_to_show! * 60 * 60 * 1000
-    );
-    const lastDate = this._lastLogbookDate || hoursToShowDate;
-    const now = new Date();
-    let newEntries: LogbookEntry[];
-
-    try {
-      [newEntries] = await Promise.all([
-        getLogbookData(
-          this.hass,
-          lastDate.toISOString(),
-          now.toISOString(),
-          this._configEntities!.map((entity) => entity.entity).toString()
-        ),
-        this._fetchUserPromise,
-      ]);
-    } catch (err: any) {
-      this._error = err.message;
-      return;
-    }
-
-    const logbookEntries = this._logbookEntries
-      ? [...newEntries, ...this._logbookEntries]
-      : newEntries;
-
-    this._logbookEntries = logbookEntries.filter(
-      (logEntry) => new Date(logEntry.when * 1000) > hoursToShowDate
-    );
-
-    this._lastLogbookDate = now;
-  }
-
-  private async _fetchUserNames() {
-    const userIdToName = {};
-
-    // Start loading users
-    const userProm = this.hass.user?.is_admin && fetchUsers(this.hass);
-
-    // Process persons
-    Object.values(this.hass!.states).forEach((entity) => {
-      if (
-        entity.attributes.user_id &&
-        computeStateDomain(entity) === "person"
-      ) {
-        this._userIdToName[entity.attributes.user_id] =
-          entity.attributes.friendly_name;
-      }
-    });
-
-    // Process users
-    if (userProm) {
-      const users = await userProm;
-      for (const user of users) {
-        if (!(user.id in userIdToName)) {
-          userIdToName[user.id] = user.name;
-        }
-      }
-    }
-
-    this._userIdToName = userIdToName;
   }
 
   static get styles(): CSSResultGroup {
@@ -317,20 +148,9 @@ export class HuiLogbookCard extends LitElement implements LovelaceCard {
           padding-top: 16px;
         }
 
-        .no-entries {
-          text-align: center;
-          padding: 16px;
-          color: var(--secondary-text-color);
-        }
-
         ha-logbook {
           height: 385px;
           display: block;
-        }
-
-        ha-circular-progress {
-          display: flex;
-          justify-content: center;
         }
       `,
     ];
